refactor(test): move local force subgraph oracle into test helpers
Drop the reintroduced production Python API and keep the reference
implementation under tests/unit/grapher/ for parity checks against the
HTML viewer's localForceSubgraph logic.
